==> R/00_all.qmd <==
---
title: "00_all"
author: 
  - "Asta Zeuner - s203544 - azeuner"
  - "Astrid Ginnerup - s203523 - AGinnerup"
  - "Emma Danø - s203564 - EDanoe"
  - "Andrea Kristensen - s193769 - AndreaMTK"
  - "Yayi Wang - s243554 - Yayi0117"
format: 
  html: 
    embed-resources: true
editor: visual
execute: 
  message: false
  warning: false
---

# Analysis of Head and Neck Squamous Cell Carcinomas Data

```{r}
#| label: "load-libraries-00" 
#| echo: TRUE
#| eval: TRUE 
#| output: FALSE 
#| include: FALSE 

library("here")

```

```{r}
#| label: "creating-result-folder"
#| include: false

# Creating results folder

results_dir <- here("results/")

if( !dir.exists(results_dir) ){
  dir.create(path = results_dir)
}

# Creating images folder

images_dir <- here("results/images/")

if( !dir.exists(images_dir) ){
  dir.create(path = images_dir)
}

```

## Loading the Dataset

```{r}
#| include: false

quarto::quarto_render(input = "01_load.qmd",
                      output_format = "html")

file.rename(from = "01_load.html",
            to = "../results/01_load.html")

```

{{< include 01_load.qmd >}}

# Cleaning the data

```{r}
#| include: false

quarto::quarto_render(input = "02_clean.qmd",
                      output_format = "html")

file.rename(from = "02_clean.html",
            to = "../results/02_clean.html")
```

{{< include 02_clean.qmd >}}

# Augmenting the data

```{r}
#| include: false

quarto::quarto_render(input = "03_augment.qmd",
                      output_format = "html")

file.rename(from = "03_augment.html",
            to = "../results/03_augment.html")
```

{{< include 03_augment.qmd >}}

# Description

```{r}
#| include: false

quarto::quarto_render(input = "04_describe.qmd",
                      output_format = "html")

file.rename(from = "04_describe.html",
            to = "../results/04_describe.html")
```

{{< include 04_describe.qmd >}}

# Analysis: Heatmap most up and down regulated proteins

```{r}
#| include: false

quarto::quarto_render(input = "../R/05_analysis_heatmap.qmd",
                      output_format = "html")

file.rename(from = "../R/05_analysis_heatmap.html",
            to = "../results/05_analysis_heatmap.html")
```

{{< include 05_analysis_heatmap.qmd >}}

# Analysis: Linear model on vital status

```{r}
#| include: false

quarto::quarto_render(input = "../R/06_analysis_lm_vital.qmd",
                      output_format = "html")

file.rename(from = "../R/06_analysis_lm_vital.html",
            to = "../results/06_analysis_lm_vital.html")
```

{{< include 06_analysis_lm_vital.qmd >}}

# Analysis: Visual

```{r}
#| include: false

quarto::quarto_render(input = "../R/07_analysis_visual.qmd",
                      output_format = "html")

file.rename(from = "../R/07_analysis_visual.html",
            to = "../results/07_analysis_visual.html")
```

{{< include 07_analysis_visual.qmd >}}

# Analysis: PCA on protein expression

```{r}
#| include: false

quarto::quarto_render(input = "../R/08_analysis_PCA.qmd",
                      output_format = "html")

file.rename(from = "../R/08_analysis_PCA.html",
            to = "../results/08_analysis_PCA.html")

```

{{< include 08_analysis_PCA.qmd >}}

# Analysis: Linear model on alcohol history

```{r}
#| include: false

quarto::quarto_render(input = "../R/09_analysis_lm_alcohol.qmd",
                      output_format = "html")

file.rename(from = "../R/09_analysis_lm_alcohol.html",
            to = "../results/09_analysis_lm_alcohol.html")


```

{{< include 09_analysis_lm_alcohol.qmd >}}

## Full report

```{r}
#| include: false

file.rename(from = "00_all.html",
            to = "../results/00_all.html")
```
