==> R/01_load.qmd <==
---
title: "01_load"
author: 
  - "Asta Zeuner - s203544 - azeuner"
  - "Astrid Ginnerup - s203523 - AGinnerup"
  - "Emma Danø - s203564 - EDanoe"
  - "Andrea Kristensen - s193769 - AndreaMTK"
  - "Yayi Wang - s243554 - Yayi0117"
format: 
  html: 
    embed-resources: true
editor: visual
execute: 
  message: false
  warning: false
---

## Load libraries

```{r}
#| label: "load-libraries-01" 
#| echo: TRUE
#| eval: TRUE 
#| output: FALSE  

library("tidyverse")
library("table1")
library("here")
```

## Download data

The data is from UCSC Xena. The data is from a cohort study, GDC TGGA Head and Neck Cancer (HNSC).

Code for downloading the datasets we will be working with and saving in data/\_raw/ folder.

```{r}
#| label: "Download-protein-data-01"

raw_dir <- here("data/_raw/")
data_file <- "TCGA-HNSC.protein.tsv.gz"
data_loc <- "https://gdc-hub.s3.us-east-1.amazonaws.com/download/"

if( !dir.exists(raw_dir) ){
  dir.create(path = raw_dir)
}
if( !file.exists(str_c(raw_dir, data_file)) ){
  download.file(
    url = str_c(data_loc, data_file),
    destfile = str_c(raw_dir, data_file))
}

```

```{r}
#| label: "Download-survival-data-01"

raw_dir <- here("data/_raw/")
data_file <- "TCGA-HNSC.survival.tsv.gz"
data_loc <- "https://gdc-hub.s3.us-east-1.amazonaws.com/download/"

if( !dir.exists(raw_dir) ){
  dir.create(path = raw_dir)
}
if( !file.exists(str_c(raw_dir, data_file)) ){
  download.file(
    url = str_c(data_loc, data_file),
    destfile = str_c(raw_dir, data_file))
}

```

```{r}
#| label: "Download-clinical-data-01"

raw_dir <- here("data/_raw/")
data_file <- "TCGA-HNSC.clinical.tsv.gz"
data_loc <- "https://gdc-hub.s3.us-east-1.amazonaws.com/download/"

if( !dir.exists(raw_dir) ){
  dir.create(path = raw_dir)
}
if( !file.exists(str_c(raw_dir, data_file)) ){
  download.file(
    url = str_c(data_loc, data_file),
    destfile = str_c(raw_dir, data_file))
}

```

## Load data

We are downloading 3 files form the Head and Neck Cancer cohort:

-   **Protein expression data:** this is a dataset including the expression levels for the different proteins in the tissue samples, for the different patients. The expression levels are obtained using RPPA, with healthy tissue samples as control. Thus the expression levels are normalized values compared to a healthy sample.

-   **Survival data**: this is a dataset including the survival status of the patients, including the number of days they survived after diagnosis, if they are not dead it is the number of days from diagnosis till last chek up.

-   **Clinical data:** this is meta data about patients in the study, including age, alcohol history etc

read_tsv is used to load data files from data/\_raw and saving the files in objects.

```{r}
#| label: "load-data-01" 

protein_data <- read_tsv(file = here("data/_raw/TCGA-HNSC.protein.tsv.gz"))
survival_data <- read_tsv(file = here("data/_raw/TCGA-HNSC.survival.tsv.gz"))
clinical_data <- read_tsv(file = here("data/_raw/TCGA-HNSC.clinical.tsv.gz"))

```

## Join meta data: Clinical and survival data

The clinical_and survival_data are joined using inner_join() to create a meta_data set. The joined meta_data is saved as a .tsv.gz file in data folder.

```{r}
#| label: "join-clinical-and-survival-data-01" 

meta_data <- inner_join(clinical_data, survival_data, by = c("sample" = "sample"))

data_dir <- here("data/")

write_tsv(x = meta_data, 
          file = str_c(data_dir, 
                       "01_meta_data_load.tsv.gz"))

```

## Joining meta_data and protein expression

The protein_data is pivoted longer using pivot_longer()

```{r}
#| label: "protein_data-pivot-longer-01"

peptide_long <- protein_data |> 
                pivot_longer(cols = -1, 
                             names_to = "sample", 
                             values_to = "expression")

```

meta_data and protein_long datasets are then joined using inner_join(). The resulting dataset is saved as a .tsv.gz file in data folder and will be the dataset used for cleaning the data.

```{r}
#| label: "Join-the-data-to-final-01"

# Join the datasets on the 'sample' column

expression_meta_data <- inner_join(meta_data, 
                         peptide_long, 
                         by = "sample")

data_dir <- here("data/")


write_tsv( x = expression_meta_data,
           file = str_c(data_dir,
                        "01_expression_meta_data_load.tsv.gz"))

```
