==> R/07_analysis_visual.qmd <==
---
title: "07_analysis_visual"
author: 
  - "Asta Zeuner - s203544 - azeuner"
  - "Astrid Ginnerup - s203523 - AGinnerup"
  - "Emma Danø - s203564 - EDanoe"
  - "Andrea Kristensen - s193769 - AndreaMTK"
  - "Yayi Wang - s243554 - Yayi0117"
format: 
  html: 
    embed-resources: true
editor: visual
execute: 
  message: false
  warning: false
---

## **Loading libraries**

```{r}
#| label: "load-libraries-07" 
#| echo: TRUE
#| eval: TRUE 
#| output: FALSE  
#| 

library("tidyverse")
library("ggplot2")
library("here")

```

## Loading data

```{r}
#| label: "load-data-07" 

expression_meta_data_aug <- read_tsv(file = here("data/03_expression_meta_data_aug.tsv.gz"))

```

## **Exploring the survival of each gender**

## 

```{r}

#| label: "survival-between-genders-barplot-07" 

barplot_vital_status_genders <- expression_meta_data_aug |> 
  ggplot(mapping = aes(x = gender,
                       fill = vital_status)) +
  geom_bar(position = "fill") +
  scale_fill_manual(values = c("Alive" = "#eb4cbe",
                               "Dead" = "#beeb4c")) +
  labs(x = "Gender", 
       y = "Proportion",
       title = "Distribution of Gender and Vital Status",
       subtitle = "Proportional representation by Gender and Race",
       fill = "Vital Status") +
  theme_classic() +
  theme(plot.title = element_text(size = 10, hjust = 0.5, face = "bold"),
        plot.subtitle = element_text(size = 8, hjust = 0.5),
        legend.position = "right")
 
barplot_vital_status_genders  

ggsave(filename = "barplot_vital_status_genders.png",
       plot = barplot_vital_status_genders, 
       device = "png",
       path = "../results/images")


```

The barplot visualizes the proportion of the vital status distributed between the two genders. The position = fill in geom_bar computes the proportion and thus allows for direct comparison between the two genders. It does though not reveal absolute numbers. The barplot indicates that the proportion of dead patients is higher within the female patients.

## Exploring the vital status between different races

With a facet on each race, we can visually interpret any deviations in vital status between the different races and still keep the distributions on gender

```{r}

#| label: "survival-between-genders-and-races-barplot-07" 

barplot_vital_status_genders_facet_races <- expression_meta_data_aug |> 
  ggplot(mapping = aes(x = gender, fill = vital_status)) +
  geom_bar(position = "fill")+
  scale_fill_manual(values = c("Alive" = "#eb4cbe",
                               "Dead" = "#beeb4c")) +
  facet_wrap(~race) +
  labs(x = "Gender", 
       y = "Proportion",
       title = "Distribution of Gender and Vital Status Across Races",
       subtitle = "Proportional representation by gender and race",
       fill = "Vital Status") +
  theme_classic() +
  theme(plot.title = element_text(size = 10, hjust = 0.5, face = "bold"),
        plot.subtitle = element_text(size = 8, hjust = 0.5),
        strip.text = element_text(size = 5, face = "bold"))

barplot_vital_status_genders_facet_races

ggsave(filename = "barplot_vital_status_genders_facet_races.png",
       plot = barplot_vital_status_genders_facet_races, 
       device = "png",
       path = "../results/images")

```

## Exploring the effect of smoking on the vital status

Constructing a boxplot to visually and statistically interpret the relationship between "years smoked" and the vital status across the different pathological stages.

```{r}

#| label: "years-smoked-vital-status-07" 

boxplot_years_smoked_vital_status_facet_pathological <- expression_meta_data_aug |> 
  drop_na(pathologic_stage) |> 
  ggplot(mapping = aes(x = vital_status,
                       y = years_smoked)) +
  geom_boxplot(fill = "#eb4cbe",
               color = "#beeb4c") +
  facet_wrap(~pathologic_stage) +
  labs(x = "Vital Status", 
       y = "Years smoked",
       title = "Distribution of Years smoked and Vital Status Across pathological stages",
       fill = "Vital Status")+
  theme_classic() +
  theme(plot.title = element_text(size = 10, 
                                  hjust = 0.5,
                                  face = "bold"))

boxplot_years_smoked_vital_status_facet_pathological

ggsave(filename = "boxplot_years_smoked_vital_status_facet_pathological.png",
       plot = boxplot_years_smoked_vital_status_facet_pathological, 
       device = "png",
       path = "../results/images")

```

Apparently, no patients are recorded with years smoked = 0. We assume they are registered as NA although NA could also refer to the fact that the patients did not respond. We can quickly determine the number of patients with NA;

```{r}
expression_meta_data_aug |> 
  filter(is.na(years_smoked)) |>
  distinct(patient_id) |> 
  nrow()
  
```

And we thus see that there are actually more than half of the patients with instances of NA in the "years smoked". The boxplot shows that for stages II and III, the surviving patients have generally smoked more years, than the dead. For stage I and IV the trend is opposite.

## Visualizing significant proteins related to vital status

As the amount of significant differentially expressed proteins between alive and dead patients were found to be a relatively low number, we can visualize their individual expression levels of each patient.

```{r}

#| label: "expression-of-6-significant-peptides-07" 

expression_of_6_significant_pep_and_vital_status <- expression_meta_data_aug |> 
  filter(peptide_target %in% c("Notch3", "FIBRONECTIN", "MERIT40_pS29","Enolase-2","PKC-a-b-II_pT638_T641","IMP3")) |> 
  
  ggplot(mapping = aes(x = age_at_index,
                       y = expression,
                       color = vital_status
                      )) +
  geom_point( shape = 8,
              size = 1) +
  
  facet_wrap(~peptide_target) +
  
  scale_color_manual(values = c("Alive" = "#eb4cbe",
                                "Dead" = "#beeb4c")) +
  labs(x = "Age", 
       y = "Expression level",
       fill = "Vital Status",
       color = "Vital Status", 
       title = "Peptides significantly expressed between vital status") +
  theme_classic() +
  theme(plot.title = element_text(hjust = 0.5, face = "bold"),
        plot.subtitle = element_text(hjust = 0.5),
        strip.text = element_text(size = 6))

expression_of_6_significant_pep_and_vital_status

ggsave(filename = "expression_of_6_significant_pep_and_vital_status.png",
       plot = expression_of_6_significant_pep_and_vital_status, 
       device = "png",
       path = "../results/images")

```

With this plot we want to graphically/visually illustrate that some peptides expression are significantly associated with death. For example, upregulation of fibronectin is statistically significant for patiens who died. This can be a bit difficult to see from the plot. What is more evidently seen from the plot is the fact that MERIT40_pS29 is downregulated in patients who died; the blue points cluster in the lower part of the plot.

**Below is a dead end:**

Trying to think of another way to display the data in a less messy way:

```{r}

expression_of_6_significant_pep_and_vital_status <- expression_meta_data_aug |> 
  filter(peptide_target %in% c("Notch3", "FIBRONECTIN", "MERIT40_pS29","Enolase-2","PKC-a-b-II_pT638_T641","IMP3")) |> 
  
  ggplot(mapping = aes(x = age_interval_10_year,
                       y = expression,
                       fill = vital_status,
                       color = vital_status)) +
  geom_point(shape = 8,
             size = 1) +
  labs(x = "Age", 
       y = "Expression level",
       fill = "Vital Status",
       color = "Vital Status") +
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5),
        legend.title = "Vital status",
        legend.position = "right") +
  
  facet_wrap(~peptide_target) +
  theme_classic()

expression_of_6_significant_pep_and_vital_status

```
